// apps/admin-web/src/.generated/api/model/updateServiceConfigRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { InquiryConfigDto } from './inquiryConfigDto';
import type { MaintenanceConfigDto } from './maintenanceConfigDto';
import type { OperationConfigDto } from './operationConfigDto';

export interface UpdateServiceConfigRequestDto {
  operation?: OperationConfigDto;
  maintenance?: MaintenanceConfigDto;
  inquiry?: InquiryConfigDto;
}